requirements.txt
pod2gen==1.0.3
yt-dlp==2024.12.23
flask==3.1.0
gunicorn==23.0.0
httpx==0.28.1
parsel==1.9.1
dateparser==1.2.0
pydantic==2.10.4
ftfy==6.3.1
python-dotenv==1.0.1
xxhash==3.5.0
